Simplify component by separating functions
